fix: org member & project developer write back to dbt
