+++ packages/frontend/src/features/virtualView/components/CreateVirtualViewModal.tsx
@@ -14,7 +14,7 @@ import { IconInfoCircle, IconTableAlias } from '@tabler/icons-react';
 import { useCallback, type FC } from 'react';
 import { z } from 'zod';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { useGitHubRepositories } from '../../../components/UserSettings/GithubSettingsPanel';
+import { useGitIntegration } from '../../../hooks/gitIntegration/useGitIntegration';
 import useHealth from '../../../hooks/health/useHealth';
 import { useProject } from '../../../hooks/useProject';
 import { useAppSelector } from '../../sqlRunner/store/hooks';
@@ -48,11 +48,11 @@ export const CreateVirtualViewModal: FC<Props> = ({ opened, onClose }) => {
     });
 
     const { data: project } = useProject(projectUuid);
-    const { data: githubRepositories, isError } = useGitHubRepositories();
+    const { data: gitIntegration, isError } = useGitIntegration(projectUuid);
 
     const canWriteToDbtProject = !!(
         health.data?.hasGithub &&
-        githubRepositories !== undefined &&
+        gitIntegration?.enabled === true &&
         !isError &&
         project?.dbtConnection.type === DbtProjectType.GITHUB
     );
